Refactoring ColumnFiltering into folder with separate base, types, and utils
